+++ textract/parsers/ps_parser.py
@@ -4,5 +4,5 @@ from ..shell import run
 def extract(filename, **kwargs):
     """Extract text from postscript files using pstotext command.
     """
-    pipe = run('pstotext %(filename)s' % locals())
-    return pipe.stdout.read()
+    stdout, stderr = run('pstotext %(filename)s' % locals())
+    return stdout